chore: uninstall react-slot-counter
