+++ package.json
@@ -46,7 +46,6 @@
     "react-loader-spinner": "5.3.4",
     "react-markdown": "8.0.7",
     "react-router-dom": "6.23.1",
-    "react-slot-counter": "3.3.0",
     "react-toastify": "11.0.2",
     "react-topbar-progress-indicator": "4.1.1",
     "sass": "1.58.1",
