--- src/sections/homepage/metrics/Metrics.module.scss
@@ -1,29 +1,0 @@
-@import 'node_modules/@iqss/dataverse-design-system/src/lib/assets/styles/design-tokens/colors.module';
-
-.metrics {
-  width: 100%;
-
-  // This is to avoid the animation on the counter when the user has reduced motion enabled
-  @media (prefers-reduced-motion: reduce) {
-    .slot_container {
-      :global .slot-counter-item-numbers {
-        transform: none !important;
-      }
-    }
-  }
-
-  h4 {
-    padding-left: 4px;
-  }
-
-  .dataset_metrics_column {
-    margin-bottom: 1.5rem;
-    border-bottom: solid 1px $dv-border-color;
-
-    @media (min-width: 768px) {
-      margin-bottom: 0;
-      border-right: solid 1px $dv-border-color;
-      border-bottom: 0;
-    }
-  }
-}
